+++ normalize/test/query_data_types/root_array_of_strings_test.dart
@@ -23,10 +23,10 @@ void main() {
 
     test('Produces correct normalized object', () {
       final normalizedResult = {};
-      normalize(
+      normalizeOperation(
         merge: (dataId, value) =>
             (normalizedResult[dataId] ??= {}).addAll(value),
-        query: query,
+        document: query,
         data: data,
       );
 
@@ -38,8 +38,8 @@ void main() {
 
     test('Produces correct nested data object', () {
       expect(
-        denormalize(
-          query: query,
+        denormalizeOperation(
+          document: query,
           read: (dataId) => normalizedMap[dataId],
         ),
         equals(data),