+++ normalize/test/query_fragments/same_object_different_fields_test.dart
@@ -46,8 +46,8 @@ void main() {
 
     test('Produces correct nested data object', () {
       expect(
-        denormalize(
-          query: query,
+        denormalizeOperation(
+          document: query,
           read: (dataId) => sharedNormalizedMap[dataId],
         ),
         equals(sharedResponse),